[Async overflow] Can't scroll vertically while over a horizontal scroller in this content https://bugs.webkit.org/show_bug.cgi?id=210356 <rdar://problem/60523731>
Reviewed by Tim Horton.

Source/WebCore:

https://dozermapper.github.io/gitbook/documentation/customconverter.html has style
that triggers mismatched containing block and z-order layer trees, triggering the creation
of an "overflow scroll proxy node" in the scrolling tree.

If we encounter such a node in our ancestor tree walk while deciding which node to send
a wheel event too, we need to jump to the node that the proxy node is representing.

Test: fast/scrolling/mac/nested-overflow-proxy-node.html

* page/scrolling/ScrollingTree.cpp:
(WebCore::ScrollingTree::handleWheelEvent):

LayoutTests:

* fast/scrolling/mac/absolute-in-overflow-scroll.html:
* fast/scrolling/mac/nested-overflow-proxy-node-expected.txt: Added.
* fast/scrolling/mac/nested-overflow-proxy-node.html: Added.

Identifier: 223262@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@259916 268f45cc-cd09-0410-ab3c-d52691b4dbfc