REGRESSION (r259843): "Missing Plug-In" when dragging an image into Mail Compose https://bugs.webkit.org/show_bug.cgi?id=210458 <rdar://problem/61735361>
Reviewed by Wenson Hsieh.

* WebCoreSupport/WebFrameLoaderClient.mm:
(WebFrameLoaderClient::createPlugin):
r259843 changed this code from making a dictionary with a set of key->object mappings
to making a dictionary with an array of keys as its only key and an array of objects
as that key's value. Not quite the same thing.

Go back to dictionaryWithObjects:forKeys:.


Identifier: 223343@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@260043 268f45cc-cd09-0410-ab3c-d52691b4dbfc