Background images should figure into visually non empty heuristic https://bugs.webkit.org/show_bug.cgi?id=208501
Reviewed by Simon Fraser.

This makes the visually non-empty heuristic treat background images the same
as it treats regular images. This is in line with first contentful paint spec in paint timing:
https://w3c.github.io/paint-timing/.

Note that the pixel count is computed based on the image size rather than the box size, as the box size might not be known at this time.
This is equivalent to the pixel reporting done for RenderImage.

Border-images and masks are excluded, as per the spec.

* rendering/RenderBox.cpp:
(WebCore::RenderBox::imageChanged):
        Call incrementVisuallyNonEmptyPixelCountIfNeeded for background images

* rendering/RenderElement.cpp:
(WebCore::RenderElement::RenderElement):
* rendering/RenderBox.cpp:
(WebCore::RenderElement::incrementVisuallyNonEmptyPixelCountIfNeeded):
* rendering/RenderBox.h:
* rendering/RenderImage.cpp:
(WebCore::RenderImage::incrementVisuallyNonEmptyPixelCountIfNeeded): Deleted.
* rendering/RenderImage.h:
        Moved incrementVisuallyNonEmptyPixelCountIfNeeded from RenderImage to RenderElement


Identifier: 223344@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@260045 268f45cc-cd09-0410-ab3c-d52691b4dbfc