[GLIB] Fix race condition in FileMonitor implementation https://bugs.webkit.org/show_bug.cgi?id=210483
Reviewed by Adrian Perez de Castro.

This is causing flaky timeouts when running resource load statistics layout tests. The problem is that we assume
FileMonitor has the last reference of the platform monitor and it's deleted on g_object_unref(), but GLib keeps
another reference that is released later on a different thread if the monitor is still active. We just need to
ensure we cancel the monitor before calling g_object_unref().

* platform/FileMonitor.h:
* platform/glib/FileMonitorGLib.cpp:
(WebCore::FileMonitor::~FileMonitor):
(WebCore::FileMonitor::didChange):
(WebCore::FileMonitor::cancel):

Identifier: 223363@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@260067 268f45cc-cd09-0410-ab3c-d52691b4dbfc