Add a heuristic for text manipulation to treat some list items as paragraph boundaries https://bugs.webkit.org/show_bug.cgi?id=210915 <rdar://problem/61907080>
Reviewed by Megan Gardner.

Source/WebCore:

Adds a mechanism to allow text manipulation to emit an item containing the current list of text manipulation
tokens early, in the case where the paragraph content iterator crosses the boundary of an element that encloses
a paragraph. Currently, the only enclosing paragraph element will be list items that have `display: block;`,
which we can take as a hint that the text in these list items should be vended as separate items, rather than as
tokens in a single item.

This may be extended in the future to other situations by adjusting logic in `isEnclosingParagraphElement`.

Test: TextManipulation.StartTextManipulationBreaksParagraphInBetweenListItems

* editing/TextManipulationController.cpp:
(WebCore::TextManipulationController::observeParagraphs):

Tools:

Add a new API test to exercise text manipulation over several different cases of lists and list items.

* TestWebKitAPI/Tests/WebKitCocoa/TextManipulation.mm:
(TestWebKitAPI::TEST):


Identifier: 223810@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@260583 268f45cc-cd09-0410-ab3c-d52691b4dbfc