All ITP database tables should reference the ObservedDomains table on DELETE CASCADE https://bugs.webkit.org/show_bug.cgi?id=210874 <rdar://problem/62209438>
Reviewed by Brady Eidson.

Source/WebKit:

This makes it easier to delete an ITP entry from the entire database.
An entry shouldn't really exist outside of ObservedDomains anyways,
because that table holds all data about the domain, so this is a good
change to make regardless of the deletion use case.

(WebKit::needsNewCreateTableSchema):
Function to check for old CREATE TABLE queries.

* NetworkProcess/Classifier/ResourceLoadStatisticsDatabaseStore.cpp:
(WebKit::ResourceLoadStatisticsDatabaseStore::ResourceLoadStatisticsDatabaseStore):
(WebKit::ResourceLoadStatisticsDatabaseStore::enableForeignKeys):
Must enable foreign keys to use DELETE CASCADE.

(WebKit::ResourceLoadStatisticsDatabaseStore::isMigrationNecessary):
(WebKit::ResourceLoadStatisticsDatabaseStore::migrateDataToNewTablesIfNecessary):
Since this is a schema change that is executed when a table is created, we
need to check if the current schema is out of date. If so, we can
create new tables using the proper reference and migrate the data.

(WebKit::ResourceLoadStatisticsDatabaseStore::openAndDropOldDatabaseIfNecessary):
We only need to migrate the data if there were no issues with the
schema (otherwise the entire database will already have been created
with the correct CREATE TABLE queries).

* NetworkProcess/Classifier/ResourceLoadStatisticsDatabaseStore.h:

Tools:

Adds a new database file with pre-entered values and the old schema.
Adds an API test which checks if the data was migrated after
initializing the ITP database.

* TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* TestWebKitAPI/Tests/WebKitCocoa/ResourceLoadStatistics.mm:
(TEST):
* TestWebKitAPI/Tests/WebKitCocoa/incorrectCreateTableSchema.db: Added.
* TestWebKitAPI/Tests/WebKitCocoa/incorrectCreateTableSchema.db-shm: Added.
* TestWebKitAPI/Tests/WebKitCocoa/incorrectCreateTableSchema.db-wal: Added.


Identifier: 223821@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@260594 268f45cc-cd09-0410-ab3c-d52691b4dbfc