Text interaction tests should ensure view is not scroll instead of looking for a zoom change https://bugs.webkit.org/show_bug.cgi?id=211056
Reviewed by Dean Jackson.

For the text interaction sub-tests, override the scroll view's delegate so as to track
whether a scroll occurred or not instead of looking at the zoom scale. As the tests are
written looking at the zoom scale is racy. It is simpler and deterministic to detect
whether the scroll view scrolled.

The purpose of the text interaction sub-tests are to ensure that zooming to reveal
the focused element is suppressed until the interaction completes with a call to
-_didFinishTextInteractionInTextInputContext. I added some more assertions to ensure this.

* TestWebKitAPI/Tests/WebKitCocoa/RequestTextInputContext.mm:
(-[TextInteractionScrollDelegate scrollViewDidScroll:]):
(TestWebKitAPI::TEST):

Identifier: 223957@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@260756 268f45cc-cd09-0410-ab3c-d52691b4dbfc