Web Inspector: Storage: can see third-party cookies https://bugs.webkit.org/show_bug.cgi?id=211092 <rdar://problem/62469078>
Reviewed by Chris Dumez.

Source/WebKit:

Test: http/tests/inspector/page/get-cookies.html

After r259649, Web Inspector no longer incorrectly bails when attempting to get cookies if
the last resource loaded by a given frame does not have cookie access. We also need to check
whether the resource we're attempting to get cookies for is first-party or third-party so as
to reflect the current cookie policy.

* WebProcess/WebPage/WebCookieJar.cpp:
(WebKit::shouldBlockCookies):

LayoutTests:

* http/tests/inspector/page/get-cookies.html:
* http/tests/inspector/page/get-cookies-expected.txt:
* http/tests/inspector/page/resources/set-cookie.php:


Identifier: 224002@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@260807 268f45cc-cd09-0410-ab3c-d52691b4dbfc