AsyncFromSyncIterator methods should not pass absent values https://bugs.webkit.org/show_bug.cgi?id=211147
Reviewed by Ross Kirsling.

JSTests:

* test262/expectations.yaml: Mark 4 test cases as passing.

Source/JavaScriptCore:

This patch implements minor spec change [1] to match async and sync iteration
from the perspective of userland `next` and `return` iterator methods.
`throw` method always receives an argument, yet we align with others to be
consistent and future-proof.

This change is already implemented in SpiderMonkey.

[1]: https://github.com/tc39/ecma262/pull/1776

* builtins/AsyncFromSyncIteratorPrototype.js:


Identifier: 224105@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@260915 268f45cc-cd09-0410-ab3c-d52691b4dbfc