TestWebKitAPI.WebKit.DefaultDisplayName and TestWebKitAPI.WebKit.CustomDisplayName are flaky failures https://bugs.webkit.org/show_bug.cgi?id=209993 <rdar://problem/61282602>
Patch by Alex Christensen <achristensen@webkit.org> on 2020-04-30
Reviewed by Chris Dumez.

The display name for the web process is set on a low-priority background queue in WebProcess::updateProcessName()
so sometimes in our test it hasn't executed yet.  If it hasn't executed yet, it will soon, so try until it succeeds.

* TestWebKitAPI/Tests/WebKitCocoa/DisplayName.mm:
(TestWebKitAPI::checkUntilDisplayNameIs):
(TestWebKitAPI::TEST):

Identifier: 224154@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@260971 268f45cc-cd09-0410-ab3c-d52691b4dbfc