[gtk] isMainThread() assert when running minibrowser in debug builds. https://bugs.webkit.org/show_bug.cgi?id=211355
Reviewed by Mark Lam.

Using NeverDestroyed<const AtomString> is discouraged if it is in the non main thread. This can be quite easily wrong:
if the running thread is one of WorkerPool, then this is wrong since AtomStringTable will be destroyed every time underlying
Thread is shutdown. If this is invoked by AutomaticThread, this is also wrong due to the same reason etc. This is why
we introduced MainThreadNeverDestroyed and use it for const AtomString. This restriction found the bug that we are using
`NeverDestroyed<const AtomString>` in non main thread. We should not do that.

This patch fixes the issue by introducing TextureMapperShaderProgram::Variable instead of using AtomString. Then this code
no longer has thread affinity.

* platform/graphics/texmap/TextureMapperShaderProgram.cpp:
(WebCore::TextureMapperShaderProgram::getLocation): Deleted.
* platform/graphics/texmap/TextureMapperShaderProgram.h:

Identifier: 224250@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@261075 268f45cc-cd09-0410-ab3c-d52691b4dbfc